+++ src/a12_hub_toolbox/a12_path.py
@@ -5,7 +5,6 @@ BELIEF_FILENAME = "belief.json"
 BELIEF_OTE1_AGG_CSV_FILENAME = "belief_ote1_agg.csv"
 BELIEF_OTE1_AGG_JSON_FILENAME = "belief_ote1_agg.json"
 BUDUNIT_FILENAME = "budunit.json"
-BUD_MANDATE_FILENAME = "bud_person_mandate_ledger.json"
 CELLNODE_FILENAME = "cell.json"
 CELL_MANDATE_FILENAME = "cell_person_mandate_ledger.json"
 BELIEVERPOINT_FILENAME = "believerpoint.json"
@@ -131,19 +130,6 @@ def create_budunit_json_path(
         belief_mstr_dir, belief_label, believer_name, bud_time
     )
     return create_path(timepoint_dir, "budunit.json")
-
-
-def create_bud_person_mandate_ledger_path(
-    belief_mstr_dir: str,
-    belief_label: LabelTerm,
-    believer_name: BelieverName,
-    bud_time: int,
-) -> str:
-    """Returns path: belief_mstr_dir\\beliefs\\belief_label\\believers\\believer_name\\buds\n\\bud_time\\bud_person_mandate_ledger.json"""
-    timepoint_dir = create_bud_dir_path(
-        belief_mstr_dir, belief_label, believer_name, bud_time
-    )
-    return create_path(timepoint_dir, "bud_person_mandate_ledger.json")
 
 
 def create_believerpoint_path(
