+++ src/a12_hub_toolbox/test/_util/test_a12_path.py
@@ -13,7 +13,6 @@ from src.a12_hub_toolbox.a12_path import (
     BELIEF_OTE1_AGG_JSON_FILENAME,
     BELIEVEREVENT_FILENAME,
     BELIEVERPOINT_FILENAME,
-    BUD_MANDATE_FILENAME,
     BUDUNIT_FILENAME,
     CELL_MANDATE_FILENAME,
     CELLNODE_FILENAME,
@@ -30,7 +29,6 @@ from src.a12_hub_toolbox.a12_path import (
     create_believerevent_path,
     create_believerpoint_path,
     create_bud_dir_path,
-    create_bud_person_mandate_ledger_path,
     create_buds_dir_path,
     create_budunit_json_path,
     create_cell_dir_path,
@@ -257,29 +255,6 @@ def test_create_budunit_json_path_ReturnsObj():
     buds_dir = create_path(sue_dir, "buds")
     timepoint_dir = create_path(buds_dir, timepoint7)
     expected_bud_path_dir = create_path(timepoint_dir, BUDUNIT_FILENAME)
-    assert gen_bud_path == expected_bud_path_dir
-
-
-def test_create_bud_person_mandate_ledger_path_ReturnsObj():
-    # ESTABLISH
-    x_belief_mstr_dir = get_module_temp_dir()
-    a23_str = "amy23"
-    sue_str = "Sue"
-    timepoint7 = 7
-
-    # WHEN
-    gen_bud_path = create_bud_person_mandate_ledger_path(
-        x_belief_mstr_dir, a23_str, sue_str, timepoint7
-    )
-
-    # THEN
-    x_beliefs_dir = create_path(x_belief_mstr_dir, "beliefs")
-    amy23_dir = create_path(x_beliefs_dir, a23_str)
-    believers_dir = create_path(amy23_dir, "believers")
-    sue_dir = create_path(believers_dir, sue_str)
-    buds_dir = create_path(sue_dir, "buds")
-    timepoint_dir = create_path(buds_dir, timepoint7)
-    expected_bud_path_dir = create_path(timepoint_dir, BUD_MANDATE_FILENAME)
     assert gen_bud_path == expected_bud_path_dir
 
 
@@ -758,20 +733,6 @@ def test_create_budunit_json_path_HasDocString():
     doc_str = f"Returns path: {doc_str}"
     # WHEN / THEN
     assert LINUX_OS or inspect_getdoc(create_budunit_json_path) == doc_str
-
-
-def test_create_bud_person_mandate_ledger_path_HasDocString():
-    # ESTABLISH
-    doc_str = create_bud_person_mandate_ledger_path(
-        belief_mstr_dir="belief_mstr_dir",
-        belief_label=belief_label_str(),
-        believer_name=believer_name_str(),
-        bud_time=bud_time_str(),
-    )
-    doc_str = doc_str.replace("buds\\bud_time", "buds\n\\bud_time")
-    doc_str = f"Returns path: {doc_str}"
-    # WHEN / THEN
-    assert LINUX_OS or inspect_getdoc(create_bud_person_mandate_ledger_path) == doc_str
 
 
 def test_create_believerpoint_path_HasDocString():
