+++ src/a15_belief_logic/belief_cell.py
@@ -15,7 +15,6 @@ from src.a04_reason_logic.reason_plan import get_dict_from_factunits
 from src.a11_bud_logic.bud import BeliefLabel
 from src.a11_bud_logic.cell import CellUnit, cellunit_shop
 from src.a12_hub_toolbox.a12_path import (
-    BUD_MANDATE_FILENAME,
     CELL_MANDATE_FILENAME,
     CELLNODE_FILENAME,
     create_belief_json_path,
@@ -34,6 +33,7 @@ from src.a12_hub_toolbox.hub_tool import (
     get_believers_downhill_event_ints,
     open_believer_file,
 )
+from src.a15_belief_logic.a15_path import BUD_MANDATE_FILENAME
 from src.a15_belief_logic.belief import get_from_dict as beliefunit_get_from_dict
