+++ src/a15_belief_logic/test/test_belief_z_cell/test_bud_mandates.py
@@ -3,8 +3,10 @@ from src.a00_data_toolbox.file_toolbox import open_json, save_json
 from src.a11_bud_logic.bud import tranbook_shop
 from src.a12_hub_toolbox.a12_path import (
     create_belief_json_path,
-    create_bud_person_mandate_ledger_path as bud_mandate_path,
     create_cell_person_mandate_ledger_path as cell_mandate_path,
+)
+from src.a15_belief_logic.a15_path import (
+    create_bud_person_mandate_ledger_path as bud_mandate_path,
 )
 from src.a15_belief_logic.belief import (
     beliefunit_shop,
