+++ src/a20_world_logic/test/test_etl_pipeline/test_calc_belief_bud_person_mandate.py
@@ -11,6 +11,8 @@ from src.a12_hub_toolbox.a12_path import (
     create_belief_json_path,
     create_belief_ote1_json_path,
     create_believerevent_path,
+)
+from src.a15_belief_logic.a15_path import (
     create_bud_person_mandate_ledger_path as bud_mandate_path,
 )
 from src.a15_belief_logic.belief import (
